% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint_gibbs_1cp_v2.R
\name{cp1_gibbs_v2}
\alias{cp1_gibbs_v2}
\title{Estimate a posterior distribution of data conditional that there is one groove.}
\usage{
cp1_gibbs_v2(data, iter, start.vals.left, start.vals.right, prop_var_left,
  prop_var_right, cp_prop_var, tol_edge = 10, warmup = 5000,
  verbose = FALSE)
}
\arguments{
\item{data}{Data frame with columns "x" and "y." "x" is a column of the locations of the
observed residual values, y.}

\item{iter}{Number of interations after warmup.}

\item{start.vals.left}{Starting values for the changepoint algorithm assuming the groove is on the left.
List with elements "sigma", "l", "cp", "beta", and "intercept." "sigma" and "l"
  are 2 element vectors where the first element is for the data to the left of the changepoint.
  "cp" is the changepoint starting value. "beta" and "intercept" are the slope and intercept
  starting values for the mean of the data model to the left of the changepoint.
 which parameterize the covariance matrix.}

\item{start.vals.right}{Starting values for the changepoint algorithm assuming the groove is on the right.}

\item{prop_var_left}{The proposal variance for the random walk Metropolis algorithm assuming
that the groove is on the left.
A two element list of the proposal variance-covariance matrices for the random
walk metropolis algorithm(s). The first element is for the data to the left of the changepoint.}

\item{prop_var_right}{The proposal variance for the random walk Metropolis algorithm assuming
that the groove is on the right.}

\item{cp_prop_var}{The proposal variance for the changepoint.}

\item{tol_edge}{This parameter controls how close changepoint proposals can be to the edge of the data
before getting automatically rejected. For example, a value of 10 means that the changepoint will be
automatically rejected if the proposal is within a distance of 10 x-values from either edge.}

\item{warmup}{The number of initial iterations which serves two purposes: the first is to allow the
algorithm to wander to the area of most mass, and the second is to tune the proposal variance.}

\item{verbose}{Logical value indicating whether to print the iteration number and the parameter proposals.}
}
\value{
A named list with all of the output that the left and right changepoint functions produce
individually plus the posterior means of the left and right changepoints.
}
\description{
This function is basically a wrapper for running the left and right (one) changepoint
Gibbs algorithms. The only computation that this function does is to estimate the posterior
means of the left and right changepoint distributions.
}
