% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint_gibbs_2cp_v2.R
\name{cp2_gibbs_v2}
\alias{cp2_gibbs_v2}
\title{Estimate a posterior distribution of data conditional that there are two grooves.}
\usage{
cp2_gibbs_v2(data, iter, start.vals, prop_var, cp_prop_var,
  tol_edge = 50, tol_cp = 1000, warmup = 500, verbose = FALSE)
}
\arguments{
\item{data}{Data frame with columns "x" and "y." "x" is a column of the locations of the
observed residual values, y.}

\item{iter}{Number of interations after warmup.}

\item{start.vals}{Starting values for the changepoint algorithm.
List with elements "sigma", "l", "cp", "beta", and "intercept." "sigma" and "l"
  are 3 element vectors where the first element is for the data on the left groove.
  The second element is for the land engraved area, and the third element is for the right groove.
  "cp" is the vector of changepoint starting values. "beta" and "intercept" are two element vectors
  of the slope and intercept for the left and right groove engraved area respectively.}

\item{prop_var}{A three element list of the proposal variance-covariance matrices for the random
walk Metropolis algorithm(s). The first element is for the left groove engraved area.
The second element is for the land engraved area, and the third element is for the right engraved area.}

\item{cp_prop_var}{The proposal variance-covariance matrix for the changepoints.}

\item{tol_edge}{This parameter controls how close changepoint proposals can be to the edge of the data
before getting automatically rejected. For example, a value of 10 means that the changepoint will be
automatically rejected if either of the proposal changepoints is within a distance of 10 x-values from either edge.}

\item{tol_cp}{This parameter controls how close changepoint proposals can be to each other
before getting automatically rejected. For example, a value of 10 means that the changepoint will be
automatically rejected if either of the proposal changepoints is within a distance of 10 x-values from either each other.}

\item{warmup}{The number of initial iterations which serves two purposes: the first is to allow the
algorithm to wander to the area of most mass, and the second is to tune the proposal variance.}

\item{verbose}{Logical value indicating whether to print the iteration number and the parameter proposals.}
}
\value{
A named list containing the sampled parameters, acceptance rates for the Metropolis steps,
log likelihood values, and proposal variance for the changepoints.
}
\description{
This function runs a random walk metropolis within Gibbs algorithm to estimate the posterior distribution
of the value of the changepoints as well as the parameters fit in each multivariate normal distribution
on either side of each changepoint. The covariance matrices are based on the exponential covariance function.
This functions assumes equally spaced locations ("x" values in the "data" argument). The distribution
to the right of the right most changepoint and to the left of the left most changepoint have
means that are a linear function of the distance from the center of the data. The slope is
constrained to be negative in the left case and positive in the right case. The models fit
to the groove engraved areas are exactly the same as in the one changepoint case. Thus, this algorithm
only differs in that there are three segments of data to deal with as opposed to two.
}
