% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myimpute.R
\name{mlgp}
\alias{mlgp}
\title{Impute missing data.}
\usage{
mlgp(y, x, tol = 1e-06)
}
\arguments{
\item{y}{Numeric y vector of response values.}

\item{x}{Numeric x vector of locations used for the covariance function.}

\item{tol}{Tolerance level for the maximum likelihood procedure to fit the Gaussian process.}
}
\description{
This function performs maximum likelihood estimation to estimate the variance parameters
in a Gaussian process with a squared exponential covariance function. These parameters could then be used
in the Gaussian process used for imputation.
}
