% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_data}
\alias{raw_data}
\title{Example of an average of 2D crosscuts from the Hamby 44 data set.}
\format{A data frame with 3346 rows and two variables:
\describe{
    \item{x}{location variable}
    \item{value}{the height of the land at the given location described by x}
}}
\source{
Hamby 44
}
\usage{
raw_data
}
\description{
This data set is essentially a 2D crosscut from a land in the Hamby 44 set of bullet land scans.
}
\keyword{datasets}
