% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_cp_gibbs_v2.R
\name{variable_cp_gibbs_v2}
\alias{variable_cp_gibbs_v2}
\title{Estimate posterior distributions for the 0, 1, or 2 changepoint case.}
\usage{
variable_cp_gibbs_v2(data, iter = 50000, start.vals = NA,
  prop_var = NA, cp_prop_var = NA, tol_edge = 50, tol_cp = 1000,
  warmup = 5000, verbose = FALSE, prior_numcp = c(1/3, 1/3, 1/3))
}
\arguments{
\item{data}{Data frame with columns "x" and "y." "x" is a column of the locations of the
observed residual values, y.}

\item{iter}{Number of interations after warmup.}

\item{start.vals}{Starting values for the changepoint algorithm. This is a named list
of lists. The names of the lists should be "cp2","cp1", and "cp0". Each list posessing
one of those aforementioned names is a list of starting values identical to what would be given
if the changepoint algorithm were to be run with the corresponding number of specified changepoints.
List with elements "sigma", "l", "cp", "beta", and "intercept." "sigma" and "l"
  are 3 element vectors where the first element is for the data on the left groove.
  The second element is for the land engraved area, and the third element is for the right groove.
  "cp" is the vector of changepoint starting values. "beta" and "intercept" are two element vectors
  of the slope and intercept for the left and right groove engraved area respectively.}

\item{prop_var}{A three element list of the proposal variance-covariance matrices for the random
walk Metropolis algorithm(s). The first element is for the left groove engraved area.
The second element is for the land engraved area, and the third element is for the right engraved area.}

\item{cp_prop_var}{The proposal variance-covariance matrix for the changepoints.}

\item{tol_edge}{This parameter controls how close changepoint proposals can be to the edge of the data
before getting automatically rejected. For example, a value of 10 means that the changepoint will be
automatically rejected if either of the proposal changepoints is within a distance of 10 x-values from either edge.}

\item{tol_cp}{This parameter controls how close changepoint proposals can be to each other
before getting automatically rejected. For example, a value of 10 means that the changepoint will be
automatically rejected if either of the proposal changepoints is within a distance of 10 x-values from either each other.}

\item{warmup}{The number of initial iterations which serves two purposes: the first is to allow the
algorithm to wander to the area of most mass, and the second is to tune the proposal variance.}

\item{verbose}{Logical value indicating whether to print the iteration number and the parameter proposals.}
}
\value{
A named list containing the sampled parameters, acceptance rates for the Metropolis steps,
log likelihood values, and proposal variance for the changepoints.
}
\description{
This function runs the changepoint functions designed for the cases when there are
0, 1, or 2 changepoints. It then returns a subset of the results that are returned for
each function individually. This subset of results is enough to decide the likely number of
shoulders, the locations of the shoulders (if they exist), as well as the posterior
samples for the changepoints for minimal diagnostic use.
}
