% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myimpute.R
\name{mlgp}
\alias{mlgp}
\title{Impute missing data.}
\usage{
mlgp(y, x, tol = 1e-06)
}
\arguments{
\item{y}{Numeric y vector.}

\item{x}{Numeric x vector.}

\item{tol}{Tolerance level for the maximum likelihood procedure to fit the Gaussian process.}
}
\description{
Functions to impute missing data.
}
