% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myimpute.R
\name{myimpute}
\alias{myimpute}
\title{Impute missing data.}
\usage{
myimpute(y, x, sigma, l)
}
\arguments{
\item{y}{Numeric y vector of response values.}

\item{x}{Numeric x vector of locations used for the covariance function.}

\item{sigma}{Marginal standard deviation in the Gaussian process.}

\item{l}{Length scale parameter in the Gaussian process.}
}
\value{
A data frame with columns "x" and "y" which contain the combined observed and imputed data.
}
\description{
This function imputes missing data based on a Gaussian process regression
}
